Move filters to separate package
