+++ svgdatagrabber/filters/__init__.py
@@ -1,0 +1,6 @@
+from .closedpath import ClosedPathFilter
+from .custom import CustomFilter
+from .filterbase import FilterBase
+from .rectangle import RectangleRangeFilter
+from .segmentnumber import SegmentNumberFilter
+from .specialline import HorizontalLineFilter, VerticalLineFilter
